analyzer/requirements.txt
flask==3.0.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0